Push FastColor through the drawing stack
FastColor's implicit uint conversions are gone, so the boundary has to
be explicit. Rather than cast at every call site - which would have put
the conversion noise inside the render inner loops - the type now flows
end to end:

- FastBitmap.GetPixel returns FastColor and the redundant
  SetPixel(.., in uint) overload goes; the backing store stays uint[]
  for memalloc.
- BitmapFont.Ink/Background, SoftwareGraphics' private helpers and the
  Elite and SCR colour fields all become FastColor, so a public
  FastColor no longer down-converts on entry.
- Alpha tests read color.A != 0 instead of masking with 0xFF000000.

Two places keep uint deliberately: AssetSet's colour-budget bookkeeping
does set arithmetic and sorts with Order(), and SDLGraphics keys its
text-texture cache on the raw ARGB.

PlanetRenderer.Landscape was doing double duty - the fractal pass filled
it with heights, ColorLandscape overwrote it in place with colours, and
the renderer drew it as pixels. That cannot survive typing, since heights
are averaged arithmetically and colours are not, so FractalPlanet now
owns a private uint[,] height map and Landscape is the FastColor buffer
it always was at draw time.

The two FastColorTests covering the deleted operators now exercise
FromUInt32/ToUInt32 instead, keeping the round trip covered.

+++ src/scr/test/StuntCarRacerSharpLib.Tests/Rendering/RoadTexturesTests.cs
@@ -36,9 +36,9 @@ public class RoadTexturesTests
 
         static void AssertStrip(FastBitmap strip, in FastColor lineColour, in FastColor roadColour)
         {
-            Assert.Equal((uint)lineColour, strip.GetPixel(0, 0));
-            Assert.Equal((uint)roadColour, strip.GetPixel(strip.Width / 2, 0));
-            Assert.Equal((uint)lineColour, strip.GetPixel(strip.Width - 1, 0));
+            Assert.Equal(lineColour, strip.GetPixel(0, 0));
+            Assert.Equal(roadColour, strip.GetPixel(strip.Width / 2, 0));
+            Assert.Equal(lineColour, strip.GetPixel(strip.Width - 1, 0));
         }
     }
